% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{bd_citation}
\alias{bd_citation}
\title{Create citations for all dependencies of bdverse}
\usage{
bd_citation(citation_level = 1, type = "raw", bib_file = NA)
}
\description{
Create citations for all dependencies of bdverse
}
