% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{bd_dependencies}
\alias{bd_dependencies}
\title{List all dependencies of bdverse}
\usage{
bd_dependencies(recursive = FALSE)
}
\arguments{
\item{recursive}{Whether to return recursive rependencies as well.}
}
\value{
tibble of dependency list
}
\description{
List all dependencies of bdverse
}
\examples{

deps <- bd_dependencies(TRUE)

}
