% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{bd_packages}
\alias{bd_packages}
\title{List all packages available in bdverse}
\usage{
bd_packages()
}
\value{
tibble of packages list
}
\description{
List all packages available in bdverse
}
\examples{


}
