% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{bdverse_app}
\alias{bdverse_app}
\title{Launch bdverse App Suite Launcher}
\usage{
bdverse_app()
}
\description{
Launch bdverse App Suite Launcher
}
\examples{

if(interactive()){

# Requires rstudio 1.2
bd_launcher()

} 

}
